third_party/WebKit/LayoutTests/http/tests/serviceworker/fetch-cors-xhr.html

<!DOCTYPE html>
<title>Service Worker: CORS XHR of fetch()</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js?pipe=sub"></script>
<body>
<script>
async_test(function(t) {
    var login1 =
      test_login(t, 'http://127.0.0.1:8000', 'username1', 'password1');
    var login2 =
      test_login(t, 'http://localhost:8000', 'username2', 'password2');
    var SCOPE = 'resources/fetch-cors-xhr-iframe.html';
    var SCRIPT = 'resources/fetch-rewrite-worker.js';
    var host_info = get_host_info();

    Promise.all([login1, login2])
      .then(function() {
          return service_worker_unregister_and_register(t, SCRIPT, SCOPE);
        })
      .then(function(registration) {
          return wait_for_update(t, registration);
        })
      .then(function(sw) {
          return wait_for_state(t, sw, 'activated');
        })
      .then(function() { return with_iframe(SCOPE); })
      .then(function(frame) {
          return new Promise(function(resolve, reject) {
              var channel = new MessageChannel();
              channel.port1.onmessage = t.step_func(function(e) {
                  assert_equals(e.data.results, 'finish');
                  unload_iframe(frame);
                  service_worker_unregister_and_done(t, SCOPE);
                });
              frame.contentWindow.postMessage({},
                                              [channel.port2],
                                              host_info['HTTP_ORIGIN']);
            });
        })
      .catch(unreached_rejection(t));
  }, 'Verify CORS XHR of fetch() in a Service Worker');
</script>
</body>
